backend/requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio>=3.7.1,<4.0.0
asyncpg==0.30.0
click==8.1.8
fastapi==0.104.1
graphql-core==3.2.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
psycopg2-binary==2.9.9
pydantic>=2.4.0
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette>=0.27.0,<0.28.0
strawberry-graphql==0.216.1
typing_extensions==4.12.2
uvicorn==0.24.0
